@@ -5,7 +5,7 @@
  * See: https://www.gatsbyjs.org/docs/use-static-query/
  */
 
-import React, {Component} from "react"
+import React, {useState} from "react"
 import Landing from "./landing"
 import "./layout.css"
 import NavBar from './navBar'
@@ -18,17 +18,17 @@ const Backdrop = styled.div`
   color:#e6f4f1;
 `
 
-class Layout extends Component {
-  state = {
-    currentView:'',
-  }
-  render = () => (
+const Layout = () => {
+  
+  const [currentView, setCurrentView] = useState('')
+  
+  return (
     <>
       <Landing />
-      <NavBar />
-      <Backdrop>
-        <About />
-        <Projects />
+      <NavBar currentView={currentView} />
+      <Backdrop className='section'>
+        <About setView={setCurrentView}/>
+        <Projects setView={setCurrentView}/>
         {/* <Contact /> */}
       </Backdrop>
     </>
